♻️ Move sink modules into Sinks namespace
