+++ lib/sequin/application.ex
@@ -65,10 +65,10 @@ defmodule Sequin.Application do
       {Oban, Application.fetch_env!(:sequin, Oban)},
       {Redix, Application.fetch_env!(:redix, :start_opts)},
       Sequin.Databases.ConnectionCache,
-      Sequin.Redis.ConnectionCache,
-      Sequin.Kafka.ConnectionCache,
-      Sequin.Nats.ConnectionCache,
-      Sequin.RabbitMq.ConnectionCache,
+      Sequin.Sinks.Redis.ConnectionCache,
+      Sequin.Sinks.Kafka.ConnectionCache,
+      Sequin.Sinks.Nats.ConnectionCache,
+      Sequin.Sinks.RabbitMq.ConnectionCache,
       SequinWeb.Presence,
       Sequin.Tracer.DynamicSupervisor,
       {Cluster.Supervisor, [topologies]},
