+++ lib/sequin/sinks/nats/nats.ex
@@ -1,4 +1,4 @@
-defmodule Sequin.Nats do
+defmodule Sequin.Sinks.Nats do
   @moduledoc false
   alias Sequin.Consumers.ConsumerEvent
   alias Sequin.Consumers.ConsumerRecord
@@ -9,7 +9,7 @@ defmodule Sequin.Nats do
               :ok | {:error, Error.t()}
   @callback test_connection(NatsSink.t()) :: :ok | {:error, Error.t()}
 
-  @client Application.compile_env(:sequin, :nats_module, Sequin.Nats.Client)
+  @client Application.compile_env(:sequin, :nats_module, Sequin.Sinks.Nats.Client)
   defdelegate send_messages(sink, messages), to: @client
 
   defdelegate test_connection(sink), to: @client
