+++ lib/sequin/sinks/rabbitmq/rabbitmq.ex
@@ -1,4 +1,4 @@
-defmodule Sequin.RabbitMq do
+defmodule Sequin.Sinks.RabbitMq do
   @moduledoc false
   alias Sequin.Consumers.ConsumerEvent
   alias Sequin.Consumers.ConsumerRecord
@@ -9,7 +9,7 @@ defmodule Sequin.RabbitMq do
               :ok | {:error, Error.t()}
   @callback test_connection(RabbitMqSink.t()) :: :ok | {:error, Error.t()}
 
-  @client Application.compile_env(:sequin, :rabbitmq_module, Sequin.RabbitMq.Client)
+  @client Application.compile_env(:sequin, :rabbitmq_module, Sequin.Sinks.RabbitMq.Client)
   defdelegate send_messages(sink, messages), to: @client
 
   defdelegate test_connection(sink), to: @client
