+++ test/sequin/redis_pipeline_test.exs
@@ -13,6 +13,7 @@ defmodule Sequin.ConsumersRuntime.RedisPipelineTest do
   alias Sequin.Factory.ConsumersFactory
   alias Sequin.Factory.DatabasesFactory
   alias Sequin.Factory.ReplicationFactory
+  alias Sequin.Sinks.RedisMock
   alias Sequin.Test.Support.Models.Character
   alias Sequin.Test.Support.Models.CharacterDetailed
 
@@ -60,7 +61,7 @@ defmodule Sequin.ConsumersRuntime.RedisPipelineTest do
         |> ConsumersFactory.insert_deliverable_consumer_record!()
         |> Map.put(:data, ConsumersFactory.consumer_record_data(record: record))
 
-      Mox.expect(Sequin.RedisMock, :send_messages, fn sink, redis_messages ->
+      Mox.expect(RedisMock, :send_messages, fn sink, redis_messages ->
         send(test_pid, {:redis_request, sink, redis_messages})
         :ok
       end)
@@ -88,7 +89,7 @@ defmodule Sequin.ConsumersRuntime.RedisPipelineTest do
         |> ConsumersFactory.insert_deliverable_consumer_record!()
         |> Map.put(:data, ConsumersFactory.consumer_record_data(record: record2))
 
-      Mox.expect(Sequin.RedisMock, :send_messages, fn sink, redis_messages ->
+      Mox.expect(RedisMock, :send_messages, fn sink, redis_messages ->
         send(test_pid, {:redis_request, sink, redis_messages})
         :ok
       end)
@@ -103,7 +104,7 @@ defmodule Sequin.ConsumersRuntime.RedisPipelineTest do
 
     @tag capture_log: true
     test "failed Redis requests result in failed events", %{consumer: consumer} do
-      Mox.expect(Sequin.RedisMock, :send_messages, fn _sink, _redis_messages ->
+      Mox.expect(Sequin.Sinks.RedisMock, :send_messages, fn _sink, _redis_messages ->
         {:error, Sequin.Error.service(service: :redis, code: "batch_error", message: "Redis batch send failed")}
       end)
 
@@ -150,7 +151,7 @@ defmodule Sequin.ConsumersRuntime.RedisPipelineTest do
     test "messages are sent from postgres to Redis", %{consumer: consumer} do
       test_pid = self()
 
-      Mox.expect(Sequin.RedisMock, :send_messages, fn sink, redis_messages ->
+      Mox.expect(Sequin.Sinks.RedisMock, :send_messages, fn sink, redis_messages ->
         send(test_pid, {:redis_request, sink, redis_messages})
         :ok
       end)
